Add lua config for builtin lsp client in neovim
* Add vim/lua directory for future lua config
* Source lua from vim/after/plugin/lua.vim, executed after all other
regular vim/plugin scripts. This was because sourcing lua directly from
init.vim resulted in an issue with the built-in LSP client, where
the diagnostics virtual text highlight group was not specified because
the LSP client was initialized before the colorscheme was specified.
Executing the LSP client config after colorscheme is set fixes the
issue.